changed job buffer top and bottom to int32_t

+++ source/Engine/wv/JobSystem/JobBuffer.cpp
@@ -16,9 +16,9 @@ wv::JobBuffer::JobBuffer(  )
 
 bool wv::JobBuffer::push( Job* _pJob )
 {
-	int bottom = m_bottom.load( std::memory_order_acquire );
+	int32_t bottom = m_bottom.load( std::memory_order_acquire );
 
-	if ( bottom < static_cast<int>( m_jobs.size() ) )
+	if ( bottom < static_cast<int32_t>( m_jobs.size() ) )
 	{
 		m_jobs[ bottom ] = _pJob;
 		m_bottom.store( bottom + 1, std::memory_order_release );
@@ -33,11 +33,11 @@ bool wv::JobBuffer::push( Job* _pJob )
 
 wv::Job* wv::JobBuffer::pop()
 {
-	long b = m_bottom.load( std::memory_order::acquire );
-	b = wv::Math::max<size_t>( 0, b - 1 );
+	int32_t b = m_bottom.load( std::memory_order::acquire );
+	b = wv::Math::max<int32_t>( 0, b - 1 );
 	m_bottom.store( b, std::memory_order::release );
 
-	long t = m_top.load( std::memory_order::acquire );
+	int32_t t = m_top.load( std::memory_order::acquire );
 
 	if ( t <= b )
 	{
@@ -45,8 +45,8 @@ wv::Job* wv::JobBuffer::pop()
 		if ( t != b )
 			return job;
 		
-		if ( !m_top.compare_exchange_weak( t, t + 1, std::memory_order::acq_rel ) ) // failed race against steal operation
-			job = nullptr;
+		if ( !m_top.compare_exchange_weak( t, t + 1, std::memory_order::acq_rel ) )
+			job = nullptr; // failed race against steal
 		
 		m_bottom.store( t + 1, std::memory_order::release );
 		return job;
@@ -61,17 +61,16 @@ wv::Job* wv::JobBuffer::pop()
 
 wv::Job* wv::JobBuffer::steal()
 {
-	long t = m_top.load( std::memory_order::acquire ) ;
+	int32_t t = m_top.load( std::memory_order::acquire ) ;
 	std::atomic_signal_fence( std::memory_order_seq_cst );
-	long b = m_bottom.load( std::memory_order::acquire );
+	int32_t b = m_bottom.load( std::memory_order::acquire );
 
 	if ( t < b )
 	{
 		Job* job = m_jobs[ t % g_NUM_JOBS ];
 		
-		// a concurrent steal or pop operation removed an element from the deque in the meantime.
-		if ( !m_top.compare_exchange_weak( t, t + 1, std::memory_order::acq_rel ) ) // failed race against steal operation
-			return nullptr;
+		if ( !m_top.compare_exchange_weak( t, t + 1, std::memory_order::acq_rel ) )
+			return nullptr; // failed race against steal or pop
 
 		return job;
 	}
